Create  Implement Bubble Sort
The share market price of a Product X varies over time interval of one hour inclusive of opening and closing time. The price in rupees recorded from 9 AM to 4 PM is as follows 40, 42, 44, 45, 46, 41, 47 and 45. Sort the prices in decreasing order and search for a given price in the list of recorded share prices. Implement Bubble Sort. Print the results after each iteration.
Input- 

8 
40 42 44 45 46 41 47 45 
43 

Output-

42 44 45 46 41 47 45 40
44 45 46 42 47 45 41 40
45 46 44 47 45 42 41 40
46 45 47 45 44 42 41 40
46 47 45 45 44 42 41 40
47 46 45 45 44 42 41 40
47 46 45 45 44 42 41 40
43 -1