C++ and graphics functions
Snake game is one of the most renowned games accessible on a wide range of gadgets and deals with each working framework. Snakes can move toward each path in this game, e.g., left, right, up, and down; in the wake of taking the food, the length of the snake increments. Food of the Snake will be created after some time