binary tree transversal
Let T be a given Binary Tree (not necessarily complete / Full / Perfect) with k elements. The elements are arranged in T such that an In-Order traversal will print the questions in Non-decreasing order of the value of the elements. Write a Program that will Re-arrange the elements in the tree such that, upon performing a Post-Order Traversal, the elements should be output in Non-decreasing order. Also, print the root node of the tree before and after re-arrangement. Your program should work for any input and be generic. You should not perform sorting on the elements.