For issue #1385
Kadane's Algorithm is used to find the largest sum contiguous subarray. The idea of Kadane’s algorithm is to maintain a variable max_ending_here that stores the maximum sum of contiguous subarray ending at the current index and a variable max_so_far stores the maximum sum of contiguous subarray found so far, every time there is a positive-sum value in max_ending_here compare it with max_so_far and update max_so_far if it is greater than max_so_far.

Please merge it.