Create kmex.cpp
You are given an array 
A
A containing 
N
N integers. Find if it is possible to choose exactly 
M
M elements from the array such that the MEX of the chosen elements is exactly 
K
K.
Recall that the MEX of an array is the smallest non-negative integer that does not belong to the array. For example, the MEX of 
[
2
,
2
,
1
]
[2,2,1] is 
0
0 because 
0
0 does not belong to the array, the MEX of 
[
3
,
1
,
0
,
1
]
[3,1,0,1] is 
2
2 because 
0
0 and 
1
1 belong to the array, but 
2
2 does not.