Create Paintingthewalls.java
Problem: 🦯
The problem you've described is a dynamic programming problem where you need to find the minimum amount of money required to paint a series of walls using two painters - a paid painter and a free painter. The paid painter has a cost associated with each wall and takes a certain amount of time to paint each wall, while the free painter can paint any wall in 1 unit of time at no cost, but can only be used when the paid painter is busy.