Add files via upload
HEY, Sir this is my code to check whether given strings are anagram or not. Anagram strings are those who have exact same number of characters and all the letters among them are common to each of the string. This is the code for it pls do accept and merge it